Support both model and encoder/decoder pair at the same time; optional cuda
The decoder/translator script used to accept only a Model or an Encoder/Decoder pair
making it tedious to use. It can now work without having to change the required
parameters.

Due to experiments cuda is not always available so make the classifier and decoder
translators work with and without cuda.